staging/requirements.txt
python-dotenv
pyyaml
psycopg2
pandas
sqlalchemy